server/requirements.txt
anyio==4.9.0
asyncpg==0.29.0
alembic==1.10.4
authlib==1.3.1
celery==5.2.7
fastapi==0.95.1
fastapi_jwt==0.3.0
flower==2.0.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-language==2.13.4
httpx==0.27.0
openai==1.79.0
python-jose==3.3.0
psycopg2-binary==2.9.6
pydantic==1.10.8
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.6
pytest==8.3.2
pytest-asyncio==0.26.0
python-dotenv==1.0.1
redis==5.0.8
sqlalchemy==2.0.41
uvicorn==0.22.0


